backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.34.1
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.1

# AI/LLM Dependencies
openai==1.98.0
langchain==0.3.27
langchain-openai==0.3.28
langchain-anthropic==0.3.18
langchain-community==0.3.27
langchain-core==0.3.72

# MCP Client Dependencies  
mcp-use==1.3.8
websockets==15.0.1

# Additional utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
python-json-logger==2.0.7
PyYAML==6.0.2